Extract TestSceneContextMenu
